+++ core/src/test/java/net/kemitix/naolo/entities/JPAActivityListenerTest.java
@@ -6,47 +6,47 @@ import org.junit.jupiter.api.Test;
 class JPAActivityListenerTest {
 
     private final JPAActivityListener listener = new JPAActivityListener();
-    private final Owner owner = new Owner();
+    private final Object object = new Object();
 
     @Test
     @DisplayName("@PrePersist")
     public void prePersist() {
-        listener.prePersist(owner);
+        listener.prePersist(object);
     }
 
     @Test
     @DisplayName("@PostPersist")
     public void postPersist() {
-        listener.postPersist(owner);
+        listener.postPersist(object);
     }
 
     @Test
     @DisplayName("@PreUpdate")
     public void preUpdate() {
-        listener.preUpdate(owner);
+        listener.preUpdate(object);
     }
 
     @Test
     @DisplayName("@PostUpdate")
     public void postUpdate() {
-        listener.postUpdate(owner);
+        listener.postUpdate(object);
     }
 
     @Test
     @DisplayName("@PreRemove")
     public void preRemove() {
-        listener.preRemove(owner);
+        listener.preRemove(object);
     }
 
     @Test
     @DisplayName("@PostRemove")
     public void postRemove() {
-        listener.postRemove(owner);
+        listener.postRemove(object);
     }
 
     @Test
     @DisplayName("@PostLoad")
     public void postLoad() {
-        listener.postLoad(owner);
+        listener.postLoad(object);
     }
 }
