+++ owners/src/main/java/net/kemitix/naolo/owners/OwnerResource.java
@@ -1,7 +1,7 @@
-package net.kemitix.naolo;
+package net.kemitix.naolo.owners;
 
 import lombok.extern.java.Log;
-import net.kemitix.naolo.entities.Owner;
+import net.kemitix.naolo.AbstractEntityResource;
 import net.kemitix.naolo.storage.*;
 
 import javax.enterprise.context.ApplicationScoped;
