+++ owners/src/main/java/net/kemitix/naolo/owners/UpdateOwner.java
@@ -1,7 +1,6 @@
-package net.kemitix.naolo.core.owners;
+package net.kemitix.naolo.owners;
 
 import lombok.Getter;
-import net.kemitix.naolo.entities.Owner;
 import net.kemitix.naolo.storage.EntityRepository;
 import net.kemitix.naolo.storage.UpdateEntityRequest;
 import net.kemitix.naolo.storage.UpdateEntityUseCase;
