+++ owners/src/test/java/net/kemitix/naolo/owners/OwnerRepositoryTest.java
@@ -1,6 +1,5 @@
-package net.kemitix.naolo.storage;
+package net.kemitix.naolo.owners;
 
-import net.kemitix.naolo.entities.Owner;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
