+++ owners/src/test/java/net/kemitix/naolo/owners/RemoveOwnerTest.java
@@ -1,6 +1,5 @@
-package net.kemitix.naolo.core.owners;
+package net.kemitix.naolo.owners;
 
-import net.kemitix.naolo.entities.Owner;
 import net.kemitix.naolo.storage.EntityRepository;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
