+++ parent/pom.xml
@@ -60,6 +60,11 @@
             </dependency>
             <dependency>
                 <groupId>net.kemitix.naolo</groupId>
+                <artifactId>naolo-owners</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>net.kemitix.naolo</groupId>
                 <artifactId>naolo-core</artifactId>
                 <version>${project.version}</version>
             </dependency>
