+++ pets/pom.xml
@@ -24,7 +24,7 @@
         </dependency>
         <dependency>
             <groupId>net.kemitix.naolo</groupId>
-            <artifactId>naolo-core</artifactId>
+            <artifactId>naolo-owners</artifactId>
         </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
