+++ pets/src/main/java/net/kemitix/naolo/pets/Pet.java
@@ -2,7 +2,7 @@ package net.kemitix.naolo.pets;
 
 import lombok.*;
 import net.kemitix.naolo.entities.JPAActivityListener;
-import net.kemitix.naolo.entities.Owner;
+import net.kemitix.naolo.owners.Owner;
 import net.kemitix.naolo.storage.HasId;
 
 import javax.json.bind.annotation.JsonbDateFormat;
