+++ war/src/test/java/net/kemitix/naolo/war/PetsIT.java
@@ -1,7 +1,7 @@
 package net.kemitix.naolo.war;
 
-import net.kemitix.naolo.OwnerResource;
-import net.kemitix.naolo.entities.Owner;
+import net.kemitix.naolo.owners.Owner;
+import net.kemitix.naolo.owners.OwnerResource;
 import net.kemitix.naolo.pets.Pet;
 import net.kemitix.naolo.pets.PetResource;
 import org.assertj.core.api.WithAssertions;
